Finance Tax Details changes for Origination..
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/Release-2.41.0@75942 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/finance/impl/FinanceDetailServiceImpl.java
@@ -4634,8 +4634,11 @@ public class FinanceDetailServiceImpl extends GenericFinanceDetailService implem
 		}
 		
 		// Delete Tax Details
-		if(financeDetail.getTaxDetail() != null){
-			getFinanceTaxDetailDAO().delete(financeDetail.getTaxDetail(), TableType.TEMP_TAB);
+		if (financeDetail.getTaxDetail() != null) {
+			FinanceTaxDetail tempTaxDetail = getFinanceTaxDetailDAO().getFinanceTaxDetail(financeDetail.getTaxDetail().getFinReference(), "_TView");
+			if (tempTaxDetail != null) {
+				getFinanceTaxDetailDAO().delete(financeDetail.getTaxDetail(), TableType.TEMP_TAB);
+			}
 		}
 
 		// Finance Main Details Deletion
